requirements.txt
PyMuPDF==1.26.4
openai>=1.40.0
pinecone>=5.0.0
tiktoken>=0.7.0
sentence-transformers>=3.0.1
numpy>=1.26.0
google-generativeai>=0.7.0
streamlit>=1.37.0
python-dotenv>=1.0.1
